requirements_notificador.txt
requests beautifulsoup4
python-telegram-bot==13.7
